Adding cl_cmdrate/cl_updaterate tuning features per issue #5
Adding cl_cmdrate and cl_updaterate tuning for issue #5, allowing
on-the-fly (relatively - required suicide built in to keypress, with
release affecting the change) changes off of bad, good, and great
(cowmanglia's comp) baseline settings, depending on which was loaded via
the controls added in the last update.  Tuning of these settings based
on the stock settings (which hold cmdrate and updaterate separate) is
not supported.  (Managing them separately would have required too many
keys.)