Fixing what is hopefully the root cause of #20 - melee taunting as medic.
Despite efforts to prevent accidental saw taunts as medic (a great way
to get dead), it would still sometimes happen, and seem to even lock -
apparently that could be achieved until this change by way of a right
click on medigun held while switching to saw.  This forces a -attack2,
thus preventing that.